Implement ATOMIC_USE_SRT_SYNC_MUTEX to replace ATOMIC_USE_POSIX_MUTEX.
